QDoc: Add a more granular representation for template declarations
When QDoc parses a project, it parses the source code to extract the
user-provided documentation and perform sanity checkings based on the
code itself on it.

When QDoc parses an "\fn" command as part of this process, it tries to
understand, based on its intermediate representation built on the
information extracted from the code-base, which "documentable element"
the "\fn" refers to.

When QDoc performs this "matching" process, it takes into consideration
only a certain amount of information.
For example, no checking is performed over the template declaration of a
callable.

Due to QDoc ignoring template declaration when "matching", certain
elements are not documentable, as they are indistinguishable.

For example, two callables that are overloaded on their specialization
will be treated as a single documentable element by QDoc, so that they
cannot both be referred to when documenting the code-base.

Currently, QDoc extracts template declarations from templated elements
directly from Clang, and stores them whole as a stringified
representation.

This representation allows QDoc to only take into consideration a whole
template declaration, as the representation lacks granularity.

Due to this representation, QDoc does not have, for example, easy access
to the information about single template parameters in a declaration.

This lack of granularity, in turn, prevents QDoc from performing more
complex operations on a template declaration.
For example, the representation would not easily allow QDoc to
understand if a documentation block mentions or not the name of a
template parameter, similar to what is done for callables parameters by
the "\a" command.

Similarly, it would only allow QDoc to perform "matching" on the whole
declaration, which might be different under the current retrieval
process when extracted from the code (where it is extracted from a
declaration) or from an "\fn" command (where it is extracted from a mock
out-of-line definition).

To simplify the implementation of features related to template
declaration, such as taking them into consideration when matching or
allowing the user to document template parameters in a sanity-checked
way, a more granular, albeit very simplified, representation for
template declarations is added to QDoc.

The representation lives in the new header file, under the root
directory for the QDoc binary source files, "template_declaration.h".

The header exports a series of types that can be used to
represent a simplified model of a template declaration, lowered to a
lossy format.

The exported types model is non-strict and aims to be lightweight, as
the representation is expected to be used as an intermediate storage for
Clang-extracted information, which is already validated.

The new representation, while currently unused, is expected to replace
the current "whole-string" representation that QDoc uses for template
declarations in the near future.

Task-number: QTBUG-118080
Task-number: QTBUG-117881
Change-Id: I6520af5ede060bd75f0ecf6b260915c80634a550
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>