QDoc: Allow RelaxedTemplateDeclaration to be rebuilt as string
When QDoc parses a project, it parses the source code to extract the
user-provided documentation and perform sanity checkings based on the
code itself on it.

When QDoc parses an "\fn" command as part of this process, it tries to
understand, based on its intermediate representation built on the
information extracted from the code-base, which "documentable element"
the "\fn" refers to.

When QDoc performs this "matching" process, it takes into consideration
only a certain amount of information.
For example, no checking is performed over the template declaration of a
callable.

Due to QDoc ignoring template declaration when "matching", certain
elements are not documentable, as they are indistinguishable.

For example, two callables that are overloaded on their specialization
will be treated as a single documentable element by QDoc, so that they
cannot both be referred to when documenting the code-base.

Currently, QDoc extracts template declarations from templated elements
directly from Clang, and stores them whole as a stringified
representation.

This representation allows QDoc to only take into consideration a whole
template declaration, as the representation lacks granularity.

Due to this representation, QDoc does not have, for example, easy access
to the information about single template parameters in a declaration.

This lack of granularity, in turn, prevents QDoc from performing more
complex operations on a template declaration.
For example, the representation would not easily allow QDoc to
understand if a documentation block mentions or not the name of a
template parameter, similar to what is done for callables parameters by
the "\a" command.

Similarly, it would only allow QDoc to perform "matching" on the whole
declaration, which might be different under the current retrieval
process when extracted from the code (where it is extracted from a
declaration) or from an "\fn" command (where it is extracted from a mock
out-of-line definition).

To simplify the implementation of features related to template
declaration, such as taking them into consideration when matching or
allowing the user to document template parameters in a sanity-checked
way, a more granular, albeit very simplified, representation for
template declarations was recently added to QDoc in
"template_declaration.h".

The representation is intended to soon replace our current whole-string
representation for template declarations.

As template declarations are shown to the user in the final output
documentation, our internal representations require us to be able to
rebuild a relatively correct human-presentable version that can be used
when the output documentation is generated.

Hence, a way to rebuild a stringified human-presentable version of the
new representation for template declarations in "template_declaration.h"
is provided.

Each relevant type in "template_declaration.h" was enhanced with a
"to_std_string" method.

The method takes care of rebuilding a human-presentable version of the
code that is modeled by the representation that can be shown to a user,
so that the new representation can be plugged in place of the current
one.

Task-number: QTBUG-118080
Task-number: QTBUG-117881
Change-Id: Ief791a7387d116ff05276ced6b0cbdcf71c1336d
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>