QDoc: Use a more granular representation for template declarations
When QDoc parses a project, it parses the source code to extract the
user-provided documentation and perform sanity checkings based on the
code itself on it.

When QDoc parses an "\fn" command as part of this process, it tries to
understand, based on its intermediate representation built on the
information extracted from the code-base, which "documentable element"
the "\fn" refers to.

When QDoc performs this "matching" process, it takes into consideration
only a certain amount of information.
For example, no checking is performed over the template declaration of a
callable.

Due to QDoc ignoring template declaration when "matching", certain
elements are not documentable, as they are indistinguishable.

For example, two callables that are overloaded on their specialization
will be treated as a single documentable element by QDoc, so that they
cannot both be referred to when documenting the code-base.

Currently, QDoc extracts template declarations from templated elements
directly from Clang, and stores them whole as a stringified
representation.

This representation allows QDoc to only take into consideration a whole
template declaration, as the representation lacks granularity.

Due to this representation, QDoc does not have, for example, easy access
to the information about single template parameters in a declaration.

This lack of granularity, in turn, prevents QDoc from performing more
complex operations on a template declaration.
For example, the representation would not easily allow QDoc to
understand if a documentation block mentions or not the name of a
template parameter, similar to what is done for callables parameters by
the "\a" command.

Similarly, it would only allow QDoc to perform "matching" on the whole
declaration, which might be different under the current retrieval
process when extracted from the code (where it is extracted from a
declaration) or from an "\fn" command (where it is extracted from a mock
out-of-line definition).

To simplify the implementation of features related to template
declarations, such as taking them into consideration when matching or
allowing the user to document template parameters in a sanity-checked
way, the more granular, albeit very simplified, representation that was
recently added in "template_declaration.h" will now be used for storing
the extracted template declarations.

Previously, QDoc would extract template declarations, where necessary,
by the usage of the mutually recursive free functions `templateDecl` and
`getTemplateParameters` in "clangcodeparser.cpp".

The two functions worked together to extract a string representation of
the whole declaration, which was later stored in a `Node`, the
intermediate representation that QDoc uses for documentable elements.

Those two functions are now removed and replaced by the
`get_template_declaration` free function, in "clangcodeparser.cpp",
which performs a similar process but extracts a template declaration as
a `RelaxedTemplateDeclaration`, a more granular representation for the
same concept.

To support the implementation of `get_template_declaration` and reduce
some code duplication, a series of free functions where added to
"clangcodeparser.cpp".

`get_expression_as_string` now takes care of retrieving a C++ expression
from the original code-base of the project. This is generally necessary
when working with certain elements that can have a default value.

The `get_default_value_initializer_as_string` overload set takes care of
extracting a stringified version of the default argument of a parameter
by dispatching to the correct concrete type some of the element of
Clang's C++ AST, which is inheritance based.

The parts of the code that previously depended on `templateDecl` and
`getTemplateParameters` are now modified to depend on usages of
`getTemplateDeclaration`.

The stored representation of a template declaration as a string in
`Node` was substituted with an optional `RelaxedTemplateDeclaration`,
to be able to store the newly used representations, keeping the general
interface for working with the stored declaration as intact as possible.

Parts of the code-base that interacted with the template declaration
that is stored in a `Node`, such as parts of the various generators that
produce the final output documentation where modified to use the new
representation, generally converting it to its stringified form to
preserve the old behavior.

The new representation is generally equivalent to the previous one with
regards to the way it is shown to the user, even if it allows for more
granularity internally.
Nonetheless, it provides for some small formatting changes.

Previously, template declarations would generally, but not always, lack
a space between the "template" keyword and the following "<" literal.

With the new representation, a the two elements are always divided by a
space.

Furthermore, QDoc now correctly acquires and shows both default argument
initializers for template parameters and information about the parameter
being a pack, while the previous representation elided those
information.

Furthermore, previously QDoc would present a mixed usage of the
"typename" and "class" keyword to present a template declaration.
With the new representation the "typename" keyword will always be
preferred.

The regression files in "tst_generatedOutput" that are touched by the
changes were updated to reflect the new state.

Task-number: QTBUG-118080
Task-number: QTBUG-117881
Change-Id: I0265b5dba81c05f5b145e926be9896a97f7e5446
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>