QDoc: Consider template declarations when matching a callable
When QDoc parses a project, it parses the source code to extract the
user-provided documentation and perform sanity checkings based on the
code itself on it.

When QDoc parses an "\fn" command as part of this process, it tries to
understand, based on its intermediate representation built on the
information extracted from the code-base, which "documentable element"
the "\fn" refers to.

When QDoc performs this "matching" process, it takes into consideration
only a certain amount of information.
For example, no checking is performed over the template declaration of a
callable.

Due to QDoc ignoring template declaration when "matching", certain
elements are not documentable, as they are indistinguishable.

For example, two callables that are overloaded on their specialization
will be treated as a single documentable element by QDoc, so that they
cannot both be referred to when documenting the code-base.

To allow QDoc to distinguish those elements, template declarations are
now taken into consideration when QDoc is matching a callable.

A new free function, `are_template_declarations_substitutable`, was
added to "template_declaration.h", the header currently exporting the
internal representation that QDoc uses for template declarations
extracted from the a code-base.

`are_template_declarations_substitutable` performs a very simple
matching procedure. This procedure does not currently follow what C++
would generally do, but it should be enough to support most cases that
should be encountered in the current documentation for the Qt project.

`findNodeForCursor`, in "clangcodeparser.cpp", the function through
which QDoc performs the general matching process for documentable
entities, was updated to make use of
`are_template_declarations_substitutable`, ensuring that the matching
process fails when two elements do not carry a substitutable template
declaration.

`findNodeForCursor` is further used to avoid storing duplicate `Node`s,
the internal representation that QDoc uses for documentable elements
extract from a code-base.
Now that it considers template declarations, QDoc will correctly store
each specialization of a callable as its own `Node`, which allows each
of those elements to be documented as their own entity.

The changes specifically address only callable elements, those
documented by an "\fn" command, and does not change how QDoc interprets
other types of templatable elements, such as classes.

Taking template declarations into consideration requires breaking
changes to the documentation of all touched templated elements in a
documentation project, as QDoc will now need those information to be
provided by the user as part of the command.
To reduce the surface of the breaking changes at the current time, all
other templatable elements were left untouched.

Task-number: QTBUG-118080
Task-number: QTBUG-117881
Change-Id: I0537dec35f351414dc63f97c7dabec506d866098
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>