qdoc: Uniform presentation of \since and \deprecated information
QDoc had a fragmented way of presenting the node (e.g. function)
status in the reference documentation; \since version was added
only to the detailed section (as a tag for the signature/section
title), but missing from the summary section. Deprecation status
(and optionally, version) was added to the summary section, but
missing from the details.

\preliminary status was only ever present in the summary section.

Further, there was no concise way of representing a status of
both \since and subsequent deprecation for the same node.

QML type reference omitted some status information that was
present for C++ types.

Introduce a nodeStatusAsString() static function in CodeMarker
that returns a string representing the node status, including
preliminary, since version, and deprecation status (and version
if defined). Use this string for both summary and details
sections.

For QML properties, drop the custom code generating this
information in favor of reusing CodeMarker::extraSynopsis(),
which now handles also QmlProperty node types.

Also in CodeMarker::extraSynopsis(), remove the trailing space
from the returned string. Let the call sites determine the
location of the delimiter, so that it correctly ends up outside
the HTML <code> element that's associated with the synopsis.

For HTML generator, modify the <code> element that displays this
information: Add classes that enable styling the element with
CSS.

Pick-to: 6.6 6.5
Fixes: QTBUG-117152
Change-Id: Ifc6ef124f6e8f74cc57a3ecbb5217c499c71dbbf
Reviewed-by: Luca Di Sera <luca.disera@qt.io>