QDoc: Reduce code duplication for default values
QDoc uses Clang APIs to extract information from the code-base of a C++
project.

As part of this process, it extracts a representation for the default
values of certain elements, such as a function parameter, so that it can
later show it to the user in the output documentation.

Due to recent changes, QDoc started extracting those default values for
multiple elements, while before it was extracting it only in one
particular case.

As part of the changes that introduced multiple usages of this form of
extraction, an extraction routine
`get_default_value_initializer_as_string`, in "clangcodeparser.cpp" was
added.

All new usages use this entry-point to perform the extraction process
for default values. Nonetheless, the original usage, the one for
extracting the default value of a function parameter declaration,
remained inline, duplicating much of the code that is abstracted for the
other cases.

Hence, use `get_default_value_initializer_as_string` to extract the
default value for function parameters to reduce the unnecessary
duplication.

`get_default_value_as_string` is an overload set, due to Clang's C++ AST
being based on inheritance, with an entry-point that dispatches a
`clang::NamedDecl`, a general base for elements that have a default
value, to the correct overload for the concrete type of the named
declaration.

A new case was added to the overload set entry point to handle a
`clang::ParmVarDecl`, the node type for a function parameter
declaration.

An overload of `get_default_value_initializer_as_string` that handles
a `clang::ParmVarDecl` was added, containing some of the original
extraction code that handles the special cases that a
`clang::ParmVarDecl` requires when obtaining its default argument.

The new overload then dispatches to `get_expression_as_string`, a
recently added free-function in "clangcodeparser.cpp" that handles
extracting a string-representation of a "clang::Expr", taking care of
performing some of the operation that were duplicated by the inline code
for function parameter extraction.

The part of `ClangVisitor::readParameterNamesAndAttributes`, the method
that takes care of extracting some of the information about a function,
including the default value for each parameter, was modified to dispatch
back to `get_default_value_initialiazer`, removing original code which
was now mostly duplicated.

Change-Id: I418b8ef30931c0dd49554c5b6b27898b413e5655
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>