QDoc: Reduce code duplication for extracting clang::Expr
QDoc uses Clang APIs to extract information from the code-base of a C++
project.

As part of this process, it sometimes need to extract an expression,
such as a default value, so that it can later show it to the user in the
output documentation.

In some recent changes, the amount of places where QDoc required the
extraction of an expression increased.

The free-function `get_expression_as_string`, in "clangcodeparser.cpp",
was introduced to handle retrieving a stringified representation of a
`clang::Expr`.

When QDoc extracts information about a function, it extracts its
noexcept status, as it is meaningful for a user.
When this is declared through a noexcept expression, such as
`noexcept(something)`, QDoc extracts the expression itself.

This extraction process replicates the process that
`get_expression_as_string` now performs and thus can be substituted by a
call to that function.

Hence, the part of `ClangVisitor::processFunction`, which takes care of
extracting various information about a function,  that takes care of
extracting the noexcept status of the function was modified to use
`get_expression_as_string` instead of duplicating the process inline.

Change-Id: I8999a90fe008b6c96dc0947833fc870be600500d
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>