Doc: Add link to Qt Design Studio in Qt Designer Manual page
This page is one of the most visited documentation pages but, according
to Google Analytics, it has very little user engagement (time spent on
site, amount of scrolling, and so on). It could be that many readers
are, in fact, looking for Qt Design Studio or Qt Quick related user
interface design.

Even though there's already a \note section after the first paragraph
that informs the user about user interface design with Qt Quick and its
benefits, add one more link to Qt Design studio and Qt Quick at the
beginning of the paragraph to allow the user to quickly access these
topics.

Fixes: QTBUG-119886
Pick-to: 6.7 6.6 6.5
Change-Id: Ie10d65537a16cf17cd2b1f6138f262ca566a7b5a
Reviewed-by: Kai Köhne <kai.koehne@qt.io>