QDoc: Remove dead code in DocParser::append
`DocParser::append` is used during the parsing of a QDoc comment-block
to store `Atom`s, an intermediate representation for the comment-block
itself, as they are extracted from the block.

`append` is an overload set and all overloads present a shared piece of
code.

When a new `Atom` is appended, if the last `Atom` was a `Code` `Atom`,
for example an `Atom` produced by QDoc encountering a "\code"/"\endcode"
pair of commands, if that `Code` `Atom` ends with an empty line, that
line is removed.

Most `Code` `Atom`s created during parsing are obtained either by
`DocParser::getCode`, the various "\code/\endcode"-like pairs, or by
using a `Quoter`, for "\quotefromfile" and related commands.

Both `getCode` and `quoted` elements will generally handle the
unnecessary line endings as part of their code path, so that when the
`Code` `Atom` is appended it should already be deprived of that kind of
format, making the special casing in `append` moot.

Hence, remove the code that takes care of this special case from all
versions of `DocParser::append`.

Change-Id: If5b5f3b06f73492fd95b3b07c0e2b347c4bb73d3
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>