Doc: Fix incorrect \l command in QDoc manual
A link command in the manual incorrectly used a mismatched parenthesis
for its first argument, instead of the correct braces, so that the
argument was read incorrectly and part of its spelling ended up in the
final output.

The command is now corrected to use braces on both sides of its first
argument.

Change-Id: Iecdd59ca2fa4d89907c4d2f1b5c583c6368f8c83
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>