QDoc: Reduce duplication in handling of "\l" command
The code taking care of parsing a "\l"ink command was separated into two
main branches.
Both branches had common code that was executed regardless of the
original branching itself.

The duplicated code was moved out of the branching to reduce
duplication.

One of the leaf of the outer branch contained an internal branch.
One of the leaf of the internal branch performs the same operations as
the other leaf of the outer branch.

Hence, the whole internal branch was moved out, following the already
moved out code and replacing the remaining branching code.

The moved out internal branch was double branching on `DocParser::isLeftBraceAhed`,
calling the method two times in succession.

The method is idempotent so that successive calls are meaningless. Hence
the double branching was flattened into a single branching operation.

Change-Id: I8e4d459c2e5d2ae4a92a7618d004a7e158e2d8ec
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>