QDoc: Condense DocParser::append
`DocParser`, the class taking care of parsing a QDoc comment-blocks, has
a private `append` method that is used all over the place to store
`Atom`s, a part of the intermediate representation for comment-blocks,
as they are read.

`DocParser::append` is an overload set that perform, except for one
specific case that is handled later, the same operation, that of
appending an `Atom` to an internal `Text`, a container for `Atom`s.

The difference between the overloads is that each one one calls a
different constructor for `Atom`, as the appended `Atom` is built in
place.

In practice, this simply wraps and duplicate the available constructors
for `Atom`, without performing any additional logic.

Hence, the `DocParser::append` overload set is removed in favor of a
new `DocParser::appendAtom` method, which performs the same appending
operation but bubbles up the `Atom` creation to the user, avoiding the
constructor duplication.

One special case of `DocParser::append` is handled slightly differently.
`DocParser::append(QString, QString)` is a special case in the overload
set as it hides some logic.

Based on the emptiness or not of the second argument, the appended atom
is a `LinkAtom`, a subclass of `Atom` that is specific to links that
have additional context for their resolution, or a base `Atom`.

This is meaningful only when parsing an "\l" command, which is in fact
its only usage. Since the logic is small and isn't really required
elsewhere it is now inlined directly in `DocParser::parse`, its single
usage location.

When appending an `Atom` to a `Text`, the `Text` will, currently,
produces a copy of the `Atom`, dispatching on overloads of `operator<<`.

It is thus required, as the original `append(QString, QString)`
achieved, that we work with the concrete type, instead of the base
`Atom` type, when appending a `LinkAtom`.

For this purpose, a second overload of `DocParser::appendAtom` for
`LinkAtom`s was included.

The interface file for `DocParser` was modified to remove the `append`
overload set, replacing it with the two `appendAtom` overloads.

Each usage of `append` in the implementation file for `DocParser` was
modified to use `appendAtom`.

Change-Id: I8de028cfb02100b4ed235e0b24fae3657abaaa2d
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>