QDoc: Remove unissuable warning
QDoc allows linking through the "\l" command.

One form of the "\l" command allows a square bracket parameter, for
example "[Qml]", that is used to restrict the search space for the link.

Internally, links that have a square bracket parameter are represented
as `LinkAtom`, a subclass of `Atom`, the general representation for
parts of a QDoc comment-block, that is modeled specifically for the
purpose of representing this special case of links.

A `LinkAtom` "resolves" the square bracket parameter.
This is done by searching for reducing the search space to a specific
`Tree`, an internal collection for elements extracted from a
code-base, or to a specific type of `Node`, an internal representation
for elements extracted from a code-base, such as a C++ `Node`.

The square bracket parameter is generally first tried to prune the
searching space by finding a `Tree`. In this form, a `Tree` instance
will generally refer to a specific module so as to limit the search
space and allow for resolving cross-module conflicting link targets.

If this fails, but the square bracket parameter is one of a few special
cases, for example "qml", it is resolved to a node type, reducing the
search space to only certain elements, for example only to elements
extracted from a QML code-base, so as to allow to resolve conflicting
link targets, such as the name of a C++ class that is the same name as a
QML type.

When no target is found, a `LinkAtom` will set an internal error member
so as to allow a warning to be reported.

`DocParser`, the class that takes care of parsing QDoc comment-blocks,
makes use of this fact when parsing an "\l" command, to report if there
was any error with the square bracket parameter, when it is present.

Nonetheless, a `LinkAtom` does not resolve its square bracket parameter
at creation time.

As it can prune the search space based on a `Tree` instance, it cannot
correctly solve the parameter QDoc does not have that information
available at the point where comment-blocks are parsed.

Indeed, the resolution of a square bracket parameter is done during the
"Generation phase" of QDoc, when the output documention is actually
produced, as QDoc will have all the required information at that point.

As a `LinkAtom` is unable to resolve the square bracket parameter at the
point where it is created, its error member will always be empty, so
that `DocParser` will never be able to emit the above-mentioned warning
when parsing an "\l" command.

Thus, the code related to issuing the warning was removed from the code
handling the parsing of an "\l" command as it is dead code.

The warning itself is currently not replaced by any equivalent warning.

A few pieces are missing to be able to construct such a warning in the
"Generation phase", where square bracket parameters are resolved.
For example, QDoc has lost the information about where the command
resides at that point, so that the warning would be unable to actually
tell the user where it should look to solve the issue.

Similarly, due to the very entrenched architecture and the timing in
which QDoc obtains certain information, it is not currently feasible to
enable issuing the warning correctly from `DocParser`.

The warning can be reinstated, if and when the code-base
is modified to address either the timing restrictions or the
availability of the location data for a link command in the "Generation
phase".

The error member that was set when resolving a square bracket parameter
was only used to check whether the warning should be issued, so that it
is now unused as a consequence of the removal.

Hence, the code related to storing, setting and retrieving the error
member for a `LinkAtom` is now removed as dead code.

Change-Id: Ic64ac0e67d3492dec93919409e1664281c5557c5
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>