qdoc: Make -outputdir cmd line option relative to current working dir
The Config instance stored a (potentially, relative) path for the
value passed to -outputdir command line option.

This value is then later used in Config::getOutputDir(). However,
when that happens, the working directory is changed already to
point to the location of the main .qdocconf file. This made the
-outputdir path also relative to the configuration file.

Resolve and store the absolute directory immediately when processing
the command line options, making the output directory relative to
the working directory at the time of executing QDoc.

The 'generatedoutput' test was abusing the way the -outputdir path
was processed: It appended a trailing / to the output directory,
which caused the HTML.nosubdirs setting to retain the original path
as the base output directory. Without the trailing separator,
HTML.nosubdirs moves the output directory one level up in the tree.
The purpose of this is to allow QDoc to output to a single directory
in a multi-module project such as Qt, where each module doc would
normally go to its own subdirectory.

As using QDir::absolutePath() now strips away any trailing/redundant
'/' or '.', such abuse is no longer possible. Adapt the
'generatedoutput' test accordingly: Set a default output subdirectory
(/html), which then gets dropped or modified in the configuration of
each individual test case.

[ChangeLog][QDoc] The path provided with the -outputdir command line
option is now relative to the current working directory when running
QDoc.

Fixes: QTBUG-77650
Change-Id: Ida1357b883ac4c5241629a3312f2864862d6111c
Reviewed-by: Paul Wicking <paul.wicking@qt.io>