Qt Designer: Fix mode switching position logic for multiple screens
The logic tries to maintain the relative form positions when switching
from top level to docked, but used the primary screen as a reference
point. This caused the forms to be out of the docked area when on a
secondary screen.

To fix this, use the widget's screen to calculate the position.

Task-number: QTBUG-120236
Pick-to: 6.7 6.6
Change-Id: I51856cdedd6a3f2c2ec73d94c0f726c04caaabca
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>