CMake: Don't pass generated ui_foo.h headers to lupdate
If the user has added foo.ui to the target sources we passed foo.ui and
ui_foo.h to lupdate. If foo.ui is id-based, ui_foo.h is generated
without context and source text information, and the created .ts file
will contain a name-less context.

We're now filtering out ui_foo.h if foo.ui is in the sources of a
target. This ensures that we'll have the correct id-based translation
generated from the information in foo.ui.

Note: If the user relies on AUTOUIC's feature that autodetects ui_foo.h
without having foo.ui in the sources, we still have the described
problem due to the lacking information in ui_foo.h. That will need to be
fixed in uic (to generate the context and source info) and lupdate (to
pick up context information, which isn't currently possible).

Fixes: QTBUG-118808
Pick-to: 6.6 6.5
Change-Id: I05a6eec3ed601908429672ee6edaf38209ebcd71
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
(cherry picked from commit d52c7a8864f2f5d2806728f1a04a033abce5b6dd)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>