QDoc: Allow brace enclosed arguments in compareswith
QDoc allows documenting the ordering relationship between a type and one
or more other types through the use of the
"\compareswith"/"\endcompareswith" command pair.

"\compareswith" takes one word to use as the comparison category, such
as "strong", and then a space separated list of words that are used to
specify the compared with types.

Due to the list being space separated, it is impossible to correctly
provide to the command a type that contains a space in its name.

For example, "unsigned long" would be always be treated as the two types
"unsigned" and "long".

To avoid the issue, the command now interprets brace-enclosed elements
as a single type, similar to how other commands in QDoc work.

For example, "{unsigned long}" will now be treated as the single type
"unsigned long".

To allow for this, "processComparesWithCommand", which takes care of
parsing the arguments passed to a "\compareswith" command, in
"docparser.cpp", was modified to parse the arguments so that
brace-enclosed arguments are treated accordingly ot the new
specification.

While QDoc already has certain procedures to parse a similar structure,
such as `DocParser::getBracedArgument`, they are coupled to the state of
the parser and are unusable in a more general situation.

In particular, due to certain structural limits and the current
implementation of the "\compareswith" command, especially its two
phase parsing to handle continued lines, the available procedures are
unusable without multiple structural changes.

Hence, an inline parser that approximates the usual braced behavior,
with some simplifications, was implemented directly in
"processComparesWithCommand".

After the command is processed and the a list of types is extracted, the
list is stored back as a string that is later extracted and used during
the generation of the output documentation set.

Previously, the list of types was joined as a space separated string,
later reconstructed during the generation by splitting on spaces.

This makes it so the original collection of words in a braced argument
that has spaces in it is lost in the translation. To avoid this, the
current behavior is kept but the list is encoded as a semi-colon
separated string and restored by splitting on semi-colons, so that
meaningful spaces are preserved.

The semi-colon character was chosen arbitrarily and has no particular
meaning.

The documentation for the "\compareswith" command was updated to reflect
the new behavior.

A simple specimen of brace-enclosed argument was added to
"generatedoutput", QDoc's end-to-end regression testing suite, under
the relevant tests for "\compareswith".

The regression files for "generatedoutput" were regenerated to respect
the changes in the regression sources.

Fixes: QTBUG-120265
Change-Id: I6872d396ae05b831478d79e66bf426f9fb353b69
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
(cherry picked from commit e66b37641bee97d0e52931ee73d1ba9ac609a7d0)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>