qdoc: Fix hyperlinks of function signatures in auto-generated lists
In generators derived from XmlGenerator, there is code that modifies
generated links for function signatures. The intent is to have the
function name act as the link, while the rest of the signature is
left outside of the <a> element as plain text.

However, this failed for signatures that do not end in closing
parenthesis but a qualifier such as 'const'. For these, the entire
signature was linked, resulting in odd and inconsistent-looking
lists of links.

Fix the regular expression that matches a function signature to
not require a match to end in a closing parenthesis.

Fixes: QTBUG-120438
Change-Id: I7e71cd2467c732ba884fba3f9781fd7b608396c2
Reviewed-by: Paul Wicking <paul.wicking@qt.io>