qhelpgenerator: localize SOURCE_DATE_EPOCH to UTC
The SOURCE_DATE_EPOCH environment variable is used as an anchor
timestamp that software builds can use when they embed timestamps
that refer to the build-time into their build artifacts.

The value of the variable is specified[1] as an integer number of
seconds since the Unix epoch (1970-01-01T00:00:00Z).

To ensure that the value is interpreted correctly, and to ensure
that the resulting timestamps embedded by qhelpgenerator cannot
drift with the build environment's timezone settings, we should
store the SOURCE_DATE_EPOCH in QDateTime instances that have a UTC
timezone.

[1] - https://reproducible-builds.org/specs/source-date-epoch/

Task-number: QTBUG-62697
Change-Id: I24a79088886adf9334cd7c3c6f723eb9f064007d
Pick-to: 6.6 6.7
Reviewed-by: Kai Köhne <kai.koehne@qt.io>