qhelpgenerator: read file modification time in UTC
During help collection construction when the SOURCE_DATE_EPOCH[1]
environment variable is set, each file modification timestamp
recorded in the TimeStampTable is clamped (capped) to a maximum of
the value specified in the environment variable.

The SOURCE_DATE_EPOCH value is an offset in seconds from the Unix
epoch, and is relative to the UTC timezone.  Therefore we can avoid
some timezone conversion operations and make it clearer in the code
that we are comparing UTC-with-UTC by adding a timezone argument[2]
of UTC when calling lastModified to retrieve the modification
timestamp for each file.  This argument is available in Qt 6.6
onwards.

This change is not intended to produce any functional change in the
behavior of the code; instead readability/comprehensibility and, to
a lesser extent, a performance benefit described by QTBUG-100349
when reading UTC-native filesystem timestamps are the two reasons
behind this change.

[1] - https://reproducible-builds.org/specs/source-date-epoch/

[2] - https://doc.qt.io/qt-6.6/qfileinfo.html#lastModified-1

Task-number: QTBUG-62697
Change-Id: Iffb1aed5ee326ed40acb6285078c6682f76750c8
Pick-to: 6.6
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>
Reviewed-by: Kai Köhne <kai.koehne@qt.io>
(cherry picked from commit b860d48aabfe43d1eb3c9c1c10bd3698d367c255)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>