CMake: FindWrapLibClang: Prevent global promotion of libzstd target
LLVM versions from 16 onward depend on libzstd and include their
own script, Findzstd.cmake, to find it. It also creates a target
that can interfere with logic in qtbase/FindWrapZSTD.cmake if it
ends up being promoted as a global target.

To avoid configuration issues, disable global promotion of the
LLVM-created libzstd target.

Pick-to: 6.7
Fixes: QTBUG-120476
Change-Id: I9159f9a1184e00c6c924bd2251ad68857d17e16c
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Kai Köhne <kai.koehne@qt.io>