QDoc: Always create a LinkAtom for an "\l" command
QDoc allows cross-referencing a documentation set or hyperlinking to an
external resource by usages of the "\l" command.

An "\l" command allows a series of different forms to customize the way
the link is presented or resolved.

In particular, an optional bracketed argument can be provided by the
user to restrict the search space for cross-referencing links.

For example, if a link to `Foo` might refer to both a C++ class and a
QML type, in a mixed language code-base, the user can force the link to
be resolved to the QML type by providing a "[qml]" argument.

Internally, QDoc, represents a link with a bracketed argument and a link
without a bracketed argument differently.

A link, is normally represented by an instance of `Atom`, one of the
general internal representations for elements extracted by a
user-provided documentation block.

If a bracketed argument is provided, the link is instead represented as
a subclass of `Atom`, `LinkAtom`, which manages the logic of parsing the
bracketed argument.

The difference between an `Atom` and a `LinkAtom` is currently
meaningful only when the link is resolved, when the output documentation
is generated, later than it is parsed.
When this happens, a `LinkAtom` resolves its bracketed argument on
request, which is used by the search routine to restrict the search
space.

Internally, a `LinkAtom` that has no bracketed argument will be treated
the same as an `Atom` by the search routine.

Since no bracketed argument is present, the relevant fields of a
`LinkAtom` will not be filled, which in turn will make it so the search
routine will be equivalent to the one for a general `Atom`.

Thus, a `LinkAtom` with no bracketed argument is equivalent to an
general `Atom`.

When an "\l" command is parsed, some branching is performed to generated
an `Atom` or a `LinkAtom` based on the presence or not of a user
provided bracketed argument.

Since `LinkAtom`s behaves equivalently to `Atom`s when they are built
without a bracketed argument, this branching is now removed in favor of
always building a `LinkAtom`, without a check on the presence or not of
a bracketed argument.

Change-Id: Ie6c25d73591ed25bc1dd3eea531531e867d54aae
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>